refactor(iota-indexer): polish logic and types
- renamed Handler trait
- removed CommonHandler struct
- removed persist_objects in IndexerStore trait
- removed duplicated CheckpointDataToCommit from the handlers module
- moved types and logic to the persist module
