+++ crates/iota-indexer/src/objects_snaphsot.rs
@@ -1,0 +1,2 @@
+// Copyright (c) 2025 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
